Isolate WireGuard code into a separate file with appropriate copyright info in header.
